Refactor the application into its own module for discovering devices on a specific destination network
